working on restructuring jobsPage
