+++ client/src/components/jobsPage/jobListing/jobDetails.tsx/details.scss
