+++ client/src/components/jobsPage/jobListing/jobDetails.tsx/map/map.scss
