+++ client/src/components/jobsPage/jobListing/jobListing.scss
