+++ client/src/components/jobsPage/jobsPage.scss
