+++ client/src/components/jobsPage/sidebar/sidebar.scss
